requirements.txt
# Core dependencies
feedparser>=6.0.0
python-telegram-bot>=20.0
apscheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# For AI summarization (optional - remove if not using Claude/GPT)
anthropic>=1.0.0
openai>=1.0.0

# For database
sqlite3 (built-in)

# For environment variables
python-dotenv>=1.0.0
